requirements.txt
Django
psycopg2-binary
pefile
Pillow
filetype
pandas
scikit-learn
joblib
